+++ RevitAddin.CommandLoader/Extensions/AppName.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using System.Reflection;
 
 namespace RevitAddin.CommandLoader.Extensions
@@ -22,7 +23,8 @@ namespace RevitAddin.CommandLoader.Extensions
             var assemblyName = assembly.GetName();
             var result = $"App: {assemblyName.Name}\n";
             result += $"Version: {assemblyName.Version.ToString(3)}\n";
-            result += $"Location: {assembly.Location}";
+            result += $"Location: {assembly.Location}\n";
+            result += $"ContextNumber: {GetContextNumber()}";
 
             return result;
         }
@@ -30,6 +32,16 @@ namespace RevitAddin.CommandLoader.Extensions
         public static string GetUri()
         {
             return "https://github.com/ricaun-io/RevitAddin.CommandLoader";
+        }
+
+        private static string GetContextNumber()
+        {
+#if NET
+            var context = System.Runtime.Loader.AssemblyLoadContext.GetLoadContext(typeof(AppName).Assembly);
+            return context.ToString().Split('#').LastOrDefault();
+#else
+            return "0";
+#endif
         }
     }
 }